aula04/pyproject.toml
[project]
name = "aula04"
version = "0.1.0"
description = ""
authors = [
    {name = "ds-borges",email = "diego97sousa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.49.1,<2.0.0)",
    "openai (>=1.108.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
